Goodbye, old ReflectionDatabase wrapper type
