+++ generate_reflection/src/api_dump.rs
@@ -1,10 +1,17 @@
 //! Interface for dealing with Roblox Studio's JSON API Dump. Isn't specific to
 //! this crate and could probably turn into a separate crate.
 
+use std::borrow::Cow;
+use std::collections::{HashMap, HashSet};
 use std::fs;
 use std::process::Command;
 
 use anyhow::Context;
+use rbx_reflection::{
+    ClassDescriptor, DataType, PropertyDescriptor, PropertyKind, PropertySerialization,
+    PropertyTag, ReflectionDatabase, Scriptability,
+};
+use rbx_types::VariantType;
 use roblox_install::RobloxStudio;
 use serde::Deserialize;
 use tempfile::tempdir;
@@ -119,4 +126,131 @@ impl Dump {
 
         Ok(dump)
     }
+
+    /// Adds all of the classes from the given API dump to the reflection
+    /// database.
+    pub fn apply(&self, database: &mut ReflectionDatabase) -> anyhow::Result<()> {
+        for dump_class in &self.classes {
+            let superclass = if dump_class.superclass == "<<<ROOT>>>" {
+                None
+            } else {
+                Some(Cow::Owned(dump_class.superclass.clone()))
+            };
+
+            let mut tags = HashSet::new();
+            for dump_tag in &dump_class.tags {
+                tags.insert(dump_tag.parse().unwrap());
+            }
+
+            let mut properties = HashMap::new();
+
+            for member in &dump_class.members {
+                if let DumpClassMember::Property(dump_property) = member {
+                    let mut tags = HashSet::new();
+                    for dump_tag in &dump_property.tags {
+                        tags.insert(dump_tag.parse().unwrap());
+                    }
+
+                    let scriptability = if tags.contains(&PropertyTag::NotScriptable) {
+                        Scriptability::None
+                    } else if tags.contains(&PropertyTag::ReadOnly) {
+                        Scriptability::Read
+                    } else {
+                        Scriptability::ReadWrite
+                    };
+
+                    let can_serialize = !tags.contains(&PropertyTag::ReadOnly)
+                        && dump_property.serialization.can_save;
+
+                    let serialization = if can_serialize {
+                        PropertySerialization::Serializes
+                    } else {
+                        PropertySerialization::DoesNotSerialize
+                    };
+
+                    // We assume that all properties are canonical by default,
+                    // since most properties are. Properties are updated by
+                    // patches later on in the database generation process.
+                    let kind = PropertyKind::Canonical { serialization };
+
+                    let type_name = &dump_property.value_type.name;
+                    let value_type = match dump_property.value_type.category {
+                        ValueCategory::Enum => DataType::Enum(type_name.clone().into()),
+                        ValueCategory::Primitive | ValueCategory::DataType => {
+                            match variant_type_from_str(type_name) {
+                                Some(variant_type) => DataType::Value(variant_type),
+                                None => continue,
+                            }
+                        }
+                        ValueCategory::Class => DataType::Value(VariantType::Ref),
+                    };
+
+                    let mut property =
+                        PropertyDescriptor::new(dump_property.name.clone(), value_type);
+                    property.scriptability = scriptability;
+                    property.tags = tags;
+
+                    // FIXME
+                    // property.kind = kind;
+
+                    properties.insert(Cow::Owned(dump_property.name.clone()), property);
+                }
+            }
+
+            let mut class = ClassDescriptor::new(dump_class.name.clone());
+            class.tags = tags;
+            class.superclass = superclass;
+            class.properties = properties;
+
+            database
+                .classes
+                .insert(Cow::Owned(dump_class.name.clone()), class);
+        }
+
+        Ok(())
+    }
+}
+
+fn variant_type_from_str(value: &str) -> Option<VariantType> {
+    Some(match value {
+        "Axes" => VariantType::Axes,
+        "BinaryString" => VariantType::BinaryString,
+        "BrickColor" => VariantType::BrickColor,
+        "CFrame" => VariantType::CFrame,
+        "Color3" => VariantType::Color3,
+        "ColorSequence" => VariantType::ColorSequence,
+        "Content" => VariantType::Content,
+        "Faces" => VariantType::Faces,
+        "Instance" => VariantType::Ref,
+        "NumberRange" => VariantType::NumberRange,
+        "NumberSequence" => VariantType::NumberSequence,
+        "PhysicalProperties" => VariantType::PhysicalProperties,
+        "Ray" => VariantType::Ray,
+        "Rect" => VariantType::Rect,
+        "Region3" => VariantType::Region3,
+        "Region3int16" => VariantType::Region3int16,
+        "UDim" => VariantType::UDim,
+        "UDim2" => VariantType::UDim2,
+        "Vector2" => VariantType::Vector2,
+        "Vector2int16" => VariantType::Vector2int16,
+        "Vector3" => VariantType::Vector3,
+        "Vector3int16" => VariantType::Vector3int16,
+        "bool" => VariantType::Bool,
+        "double" => VariantType::Float64,
+        "float" => VariantType::Float32,
+        "int" => VariantType::Int32,
+        "int64" => VariantType::Int64,
+        "string" => VariantType::String,
+
+        // ProtectedString is handled as the same as string
+        "ProtectedString" => VariantType::String,
+
+        // TweenInfo is not supported by rbx_types yet
+        "TweenInfo" => return None,
+
+        // These types are not generally implemented right now.
+        "QDir" | "QFont" => return None,
+
+        _ => panic!("Unknown type {}", value),
+    })
 }
